"Featurable" support & participatory processes plugin (#1683)
* Refactor participatory processes to a plugin

And add a new API to register new featurables. We now have an analogy
between "feature plugins" and `register_feature` and "featurable
plugins" and `register_featurable`.

* Allow featurables to define their own admins

Can't really migrate participatory process admins to work per-featurable
yet, but this allows to at least specify global admins for them.

* Enable featurable specific route constraints

This sets the example of the route constraint a featurable plugin needs
to define.

* Move models to participatory process plugin

I don't want to deal with namespacing just yet because hardcoded model
names live in DB so it will require to deal with data migrations which I
don't want to handle now.

* Migrate participatory process factories

* Move migrations to participatory processes plugin

* Move icon to processes plugin

* Move menu registration to processes plugin

* Move abilities to participatory processes plugin